refactor: extract AST form page logic
